feat(kan-39): export models into folder model update import
