+++ src/app/shared/model/stepFormResult.model.ts
@@ -1,0 +1,13 @@
+export interface StepFormResult {
+  title: string;
+  city: string | null;
+  country: string | null;
+  continent: string | null;
+  latitude: number;
+  longitude: number;
+  description: string;
+  mediaUrl: string | null;
+  startDate: string | null;
+  endDate: string | null;
+  themeId: number | null;
+}
